Converting all tests to llvm-lit tests
