--- tools/checked-c-convert/functests/ntarr/basic_inter_field.expected.c
@@ -1,23 +1,0 @@
-#include <string_checked.h>
-// This tests the propagation of constraints
-// within the fields of structure.
-typedef struct {  
-  _Ptr<int> ptr;
-  _Nt_array_ptr<char> ntptr; 
-} foo;
-
-foo obj1;
-
-_Ptr<int> func(_Ptr<int> ptr, _Nt_array_ptr<char> ntptr) {
-   obj1.ptr = ptr;
-   obj1.ntptr = strstr(ntptr, "world");
-   return ptr;
-}
-
-int main() {
-  int a;
-  _Ptr<int> b;
-  _Nt_array_ptr<char> wil;
-  a = strlen(wil);
-  b = func(&a, wil);
-}
